+++ features/include/pcl/features/impl/ppf.hpp
@@ -93,7 +93,7 @@ pcl::PPFEstimation<PointInT, PointNT, PointOutT>::computeFeature (PointCloudOut 
         }
         else
         {
-          PCL_ERROR ("[pcl::%s::computeFeature] Computing pair feature vector between points %zu and %zu went wrong.\n", getClassName ().c_str (), i, j);
+          PCL_ERROR ("[pcl::%s::computeFeature] Computing pair feature vector between points %lu and %lu went wrong.\n", getClassName ().c_str (), (unsigned long) i, (unsigned long) j);
           p.f1 = p.f2 = p.f3 = p.f4 = p.alpha_m = 0.0;
         }
       }
